+++ test/endpoints.js
@@ -1,5 +1,6 @@
-const config = require('../src/config');
-
 module.exports = {
-    tcp:'tcp://'+config.TCP_HOST+':6371'
+    tcpbinary:'tcp://localhost:6370',
+    tlsbinary:'tls://localhost:6371',
+    tcpjsonrpc:'tcp://localhost:6372',
+    tlsjsonrpc:'tls://localhost:6373'
 };