+++ test/tests/200.server.maxClients.js
@@ -1,21 +1,21 @@
 const Client = require('sloki-node-client');
-const endpoint = require('../endpoints').tcp;
 const config = require('../../src/config');
+const engine = process.env.SLOKI_SERVER_ENGINE||'tcpbinary';
+const endpoint = require('../endpoints')[engine];
 
-let maxClients = config.TCP_MAX_CLIENTS;
+const MAX_CLIENTS = config.getMaxClients(engine);
 
 require('./client')(__filename, (test, client, end) => {
     test.test('client1: getMaxClients', subtest  => {
         client.maxClients((err, result) => {
             subtest.deepEqual(err, undefined, 'method should not return an error');
-            subtest.equal(result, maxClients, 'client1: maxClients should return '+maxClients);
+            subtest.equal(result, MAX_CLIENTS, 'client1: maxClients should return '+MAX_CLIENTS);
             subtest.end();
         });
     });
 
     test.test('client1: setMaxClients', subtest  => {
-        maxClients = 'a';
-        client.maxClients({ value:maxClients }, (err) => {
+        client.maxClients({ value:'a' }, (err) => {
             const expectedError = { code: -32602, message: 'method "maxClients": property "value" should be a number, found string' };
             subtest.deepEqual(err, expectedError, 'method should an return error');
             subtest.end();
@@ -23,10 +23,9 @@ require('./client')(__filename, (test, client, end) => {
     });
 
     test.test('client1: setMaxClients', subtest  => {
-        maxClients = 1;
-        client.maxClients({ value:maxClients }, (err, result) => {
+        client.maxClients({ value:1 }, (err, result) => {
             subtest.deepEqual(err, undefined, 'method should not return an error');
-            subtest.equal(result, maxClients, 'client1: maxClients should be set to '+maxClients);
+            subtest.equal(result, 1, 'client1: maxClients should be set to 1');
             subtest.end();
         });
     });
@@ -34,16 +33,17 @@ require('./client')(__filename, (test, client, end) => {
     test.test('client1: getMaxClients', subtest  => {
         client.maxClients((err, result) => {
             subtest.deepEqual(err, undefined, 'method should not return an error');
-            subtest.equal(result, maxClients, 'client1: maxClients should be '+maxClients);
+            subtest.equal(result, 1, 'client1: maxClients should be 1');
             subtest.end();
         });
     });
 
     test.test('client2: hit maxClients', subtest => {
-        const client2 = new Client(endpoint);
+        const client2 = new Client(endpoint, { engine });
         client2
             .connect()
             .then(() => {
+                subtest.notOk('connect succeed, should not');
             })
             .catch((err) => {
                 const expectedErr = { code: -32000, message: 'Max Clients Reached' };
@@ -54,9 +54,9 @@ require('./client')(__filename, (test, client, end) => {
     });
 
     test.test('client1: restore maxClients', subtest  => {
-        client.maxClients({ value:config.TCP_MAX_CLIENTS }, (err, result) => {
+        client.maxClients({ value:MAX_CLIENTS }, (err, result) => {
             subtest.deepEqual(err, undefined, 'method should not return an error');
-            subtest.equal(result, config.TCP_MAX_CLIENTS, 'client1: maxClients should be set to '+config.TCP_MAX_CLIENTS);
+            subtest.equal(result, MAX_CLIENTS, 'client1: maxClients should be set to '+MAX_CLIENTS);
             subtest.end();
             end();
         });