+++ CHANGELOG.md
@@ -5,12 +5,13 @@
 ### Added
 * add method `versions` [#19](https://github.com/sloki-project/sloki/issues/19)
 * add alias property for all methods [#21](https://github.com/sloki-project/sloki/issues/21)
-* add custom tcp protocol [#22](https://github.com/sloki-project/sloki/issues/22)
+* add binary tcp protocol [#22](https://github.com/sloki-project/sloki/issues/22)
 * add manual garbage collector [#23](https://github.com/sloki-project/sloki/issues/23) [#24](https://github.com/sloki-project/sloki/issues/24)
 
 ### Changed
 * switch to named parameter in jsonrpc layer [#20](https://github.com/sloki-project/sloki/issues/20)
 * no response needed with jsonrpc id = -1
+* listen on multiple port per transport/protocol by default [#25](https://github.com/sloki-project/sloki/issues/25)
 
 
 ## [0.0.6](https://github.com/sloki-project/sloki/milestone/4) - 2019-02-09